requirements.txt
twilio
flask
requests
googlesearch-python
openai